+++ ffm/src/main/java/it/fantacalcio/ffm/domain/entity/Svincolo.java
@@ -20,6 +20,6 @@ public class Svincolo {
     private Operazione idOperazione;
 
     @Column(name = "PERCENTUALE", nullable = false)
-    private Byte percentuale;
+    private Integer percentuale;
 
 }
