+++ ffm/src/main/java/it/fantacalcio/ffm/domain/entity/TransazioneOperazione.java
@@ -20,7 +20,7 @@ public class TransazioneOperazione {
     private Operazione idOperazione;
 
     @Column(name = "IMPORTO", nullable = false)
-    private Byte importo;
+    private Integer importo;
 
     @Column(name = "TIPO_TRANSAZIONE", nullable = false, length = 1)
     private String tipoTransazione;
